+++ App/plugins/cordova-plugin-ionic-webview/src/ios/IONAssetHandler.m
@@ -8,6 +8,15 @@
     self.basePath = assetPath;
 }
 
+- (instancetype)initWithBasePath:(NSString *)basePath andScheme:(NSString *)scheme {
+    self = [super init];
+    if (self) {
+        _basePath = basePath;
+        _scheme = scheme;
+    }
+    return self;
+}
+
 - (void)webView:(WKWebView *)webView startURLSchemeTask:(id <WKURLSchemeTask>)urlSchemeTask
 {
     NSString * startPath = @"";
@@ -15,20 +24,26 @@
     NSString * stringToLoad = url.path;
     NSString * scheme = url.scheme;
 
-    if ([scheme isEqualToString:IONIC_SCHEME]) {
-        startPath = self.basePath;
-        if ([stringToLoad isEqualToString:@""] || [url.pathExtension isEqualToString:@""]) {
-            startPath = [startPath stringByAppendingString:@"/index.html"];
+    if ([scheme isEqualToString:self.scheme]) {
+        if ([stringToLoad hasPrefix:@"/_app_file_"]) {
+            startPath = [stringToLoad stringByReplacingOccurrencesOfString:@"/_app_file_" withString:@""];
         } else {
-            startPath = [startPath stringByAppendingString:stringToLoad];
-        }
-    } else if ([scheme isEqualToString:IONIC_FILE_SCHEME]) {
-        if (![stringToLoad isEqualToString:@""]) {
-            startPath = stringToLoad;
+            startPath = self.basePath ? self.basePath : @"";
+            if ([stringToLoad isEqualToString:@""] || [url.pathExtension isEqualToString:@""]) {
+                startPath = [startPath stringByAppendingString:@"/index.html"];
+            } else {
+                startPath = [startPath stringByAppendingString:stringToLoad];
+            }
         }
     }
-
-    NSData * data = [[NSData alloc] initWithContentsOfFile:startPath];
+    NSError * fileError = nil;
+    NSData * data = nil;
+    if ([self isMediaExtension:url.pathExtension]) {
+        data = [NSData dataWithContentsOfFile:startPath options:NSDataReadingMappedIfSafe error:&fileError];
+    }
+    if (!data || fileError) {
+        data =  [[NSData alloc] initWithContentsOfFile:startPath];
+    }
     NSInteger statusCode = 200;
     if (!data) {
         statusCode = 404;
@@ -67,7 +82,7 @@
 -(BOOL) isMediaExtension:(NSString *) pathExtension {
     NSArray * mediaExtensions = @[@"m4v", @"mov", @"mp4",
                            @"aac", @"ac3", @"aiff", @"au", @"flac", @"m4a", @"mp3", @"wav"];
-    if ([mediaExtensions containsObject:pathExtension]) {
+    if ([mediaExtensions containsObject:pathExtension.lowercaseString]) {
         return YES;
     }
     return NO;